Add the MovieDetails object and its properties.
MovieResponse will now only return the key and title of the movies being requested. A user can pull a movie from that list and request details about the movie, which will make a request to the actual resource for the movie. Annoyingly, the API returns most of the information, but a very limited amount of data for actors/roles, directors, writers, genres, etc. Requesting the resource through MovieDetails makes a new request to the URI that contains more expansive data.